e2e-coref-fork/requirements.txt
tensorflow-cpu>=1.13.1
tensorflow-hub>=0.4.0
h5py
nltk
pyhocon
scipy
sklearn
